% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_F.R
\name{lme_F}
\alias{lme_F}
\title{Estimate F statistics}
\usage{
lme_F(stats, C)
}
\arguments{
\item{stats}{Output from the \code{lme_fit_FS} function}

\item{C}{Contrast vector}
}
\value{
This function returns a list with entries F, pval, sgn, and df.
}
\description{
Estimate F statistics
}
\examples{
\dontrun{stats <- lme_fit_FS(X, Zcols, y, ni)}
\dontrun{C <- c(0, 1, 0, 0)}
\dontrun{F <- lme_F(stats, C)}
}
