% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_mass_FDR.R
\name{lme_mass_FDR}
\alias{lme_mass_FDR}
\title{Conduct FDR thresholding}
\usage{
lme_mass_FDR(p, fdr)
}
\arguments{
\item{p}{a vector of p-values}

\item{fdr}{threshold}
}
\value{
The function returns an updated threshold for uncorrected p-values that keeps the FDR.
}
\description{
Conduct FDR thresholding
}
\examples{
\dontrun{pthFDR <- lme_mass_FDR(pvals, 0.05)}
}
