% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_mass_FDR2.R
\name{lme_mass_FDR2}
\alias{lme_mass_FDR2}
\title{Perform two-stage FDR correction}
\usage{
lme_mass_FDR2(pval, sgn, maskvtx = NA, rate = 0.05, tail = 0)
}
\arguments{
\item{pval}{Vector of p-values from \code{lme_mass_F}}

\item{sgn}{Vector of signs from \code{lme_mass_F}}

\item{maskvtx}{Indices (one-based) for vertices to include in the analysis (default: all)}

\item{rate}{Alpha level (default: 0.05)}

\item{tail}{One of -1 for left-sided or 0 for two-sided or 1 for right-sided hypothesis testing (default: 0)}
}
\value{
The function returns a list with field detvtx, sided_pval, pth, and m0.
}
\description{
Perform two-stage FDR correction
}
\examples{
#'
\dontrun{F <- lme_mass_F(...)}
\dontrun{F_FDR2 <- lme_mass_FDR2(F$pval, F$sgn)}
}
