% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_mass_fit_Rgw.R
\name{lme_mass_fit_Rgw}
\alias{lme_mass_fit_Rgw}
\title{Title}
\usage{
lme_mass_fit_Rgw(
  X,
  Zcols,
  Y,
  ni,
  Th0,
  Rgs,
  Surf,
  fname = NA,
  Dtype = "euc",
  sptm = "exp",
  prs = 1,
  e = 0.1
)
}
\arguments{
\item{e}{}
}
\value{

}
\description{
Title
}
