% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_mass_fit_vw.R
\name{lme_mass_fit_vw}
\alias{lme_mass_fit_vw}
\title{Mass-univariate linear mixed-effects model fitting}
\usage{
lme_mass_fit_vw(
  X,
  Zcols,
  Y,
  ni,
  maskvtx = NA,
  prs = 1,
  e = 10^-1,
  Xrows = NA,
  numcore = 1
)
}
\arguments{
\item{X}{Design matrix}

\item{Zcols}{Vector of random effect indices in design matrix}

\item{Y}{Outcome variable}

\item{ni}{Vector indicating the repeated observations of each subject}

\item{maskvtx}{Indices (one-based) for vertices to include in the analysis (default: all)}

\item{prs}{(currently ignored)}

\item{e}{Tolerance (default: 10^-1)}

\item{Xrows}{Internal parameter (default: NA)}

\item{numcore}{Number of cores for parallel computing (default: 1)}
}
\value{
This function returns a list with the following parameters:
Bhat, CovBhat, phisqhat, Dhat, X, Zcols, invEI, Pth, Qthth, lreml.
}
\description{
Mass-univariate linear mixed-effects model fitting
}
\examples{
stats <- lme_mass_fit_vw(X, Zcols, Y, ni, numcore=1)
}
