% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_mass_rfx.R
\name{lme_mass_rfx}
\alias{lme_mass_rfx}
\title{Estimation of subject-specific random effects estimates at each vertex}
\usage{
lme_mass_rfx(stats, X, Zcols, Y, ni, maskvtx = NA)
}
\arguments{
\item{stats}{Structure array containing statistics for every voxel/vertex (generated with either lme_mass_fit_Rgw or lme_mass_fit_vw)}

\item{X}{Ordered design matrix (according to time for each subject)}

\item{Zcols}{Vector with the indices of the colums of X that are considered as random effects}

\item{Y}{Ordered data matrix (n x nv, n=total number of scans and nv=number of vertices/voxels)}

\item{ni}{Vector whose m entries are the number of repeated measures for each subject (ordered according to X)}

\item{maskvtx}{Mask's vertices (1-based). Default: NA (all vertices included)}
}
\value{
This function returns the subject-specific random effects estimates at each vertex. The output is a list of lists, with the following entries:
Rfx: Estimated subject-specific random effects matrix (m x nrfx*nv). The columns of this matrix are grouped by vertex. For example if
there are two random effects \% in the model then the first two columns contain the subject-specific random effect coefficients for the first vertex,
then the next two columns contain the subject-specific random effect coefficients for the second vertex and so on ...
nrfx: Number of random effects (length(Zcols)).
Bhat: Population-level regression coefficients in stats stacked in one matrix.
}
\description{
Estimation of subject-specific random effects estimates at each vertex
}
\examples{
\dontrun{fitRgw <- lme_mass_fit_Rgw(X, Zcols, Y, ni, fitInit$Theta0, RgGrow$Rgs, Surf)}
\dontrun{rfx <- lme_mass_rfx(fitRgw$stats, X, Zcols, Y, ni, maskvtx)}
}
