% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_openmgh.R
\name{lme_openmgh}
\alias{lme_openmgh}
\title{Read FreeSurfer image (mgh) files}
\usage{
lme_openmgh(fname)
}
\arguments{
\item{fname}{Filename of a FreeSurfer mgh file.}
}
\value{
The function returns a list with entries
x, v, ndim1, ndim2, ndim3, nframes, type, and dof.
}
\description{
Read FreeSurfer image (mgh) files
}
\examples{
mgh <- lme_openmgh(filename)
}
