% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_readlabel.R
\name{lme_readlabel}
\alias{lme_readlabel}
\title{Read FreeSurfer label files}
\usage{
lme_readlabel(fname)
}
\arguments{
\item{fname}{Filename of a FreeSurfer label file.}
}
\value{
The function returns a nvertices-by-5 matrix, where each column means:
(1) vertex number, (2-4) xyz at each vertex, (5) label value. The
vertex number is 0-based.
}
\description{
Read FreeSurfer label files
}
\examples{
\dontrun{labels <- lme_readlabel(filename)}
}
