% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_readsurf.R
\name{lme_readsurf}
\alias{lme_readsurf}
\title{Read FreeSurfer surface files}
\usage{
lme_readsurf(fname)
}
\arguments{
\item{fname}{Filename of a FreeSurfer surface file.}
}
\value{
The function returns a list with entries vertices and faces.
}
\description{
Read FreeSurfer surface files
}
\examples{
surf <- lme_readsurf(filename)
}
