% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_savemgh.R
\name{lme_savemgh}
\alias{lme_savemgh}
\title{Write FreeSurfer image (mgh) files}
\usage{
lme_savemgh(vol, fname)
}
\arguments{
\item{vol}{List with fields "ndim1", "ndim2", "ndim3", "nframes", which need to contain image dimensions, and "x", which contains the image data.}

\item{fname}{Filename}
}
\description{
Write FreeSurfer image (mgh) files
}
\examples{
\dontrun{vol <- NULL}
\dontrun{vol$ndim1 <- 256}
\dontrun{vol$ndim1 <- 256}
\dontrun{vol$ndim1 <- 160}
\dontrun{vol$ndim1 <- 1}
\dontrun{vol$x <- array(1, dim=c(256, 256, 160, 1))}
\dontrun{lme_savemgh(vol, FILENAME)}
}
