% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_mass_fit_Rgw.R
\name{lme_mass_fit_Rgw}
\alias{lme_mass_fit_Rgw}
\title{Title}
\usage{
lme_mass_fit_Rgw(
  X,
  Zcols,
  Y,
  ni,
  Th0,
  Rgs,
  Surf,
  fname = NA,
  Dtype = "euc",
  sptm = "exp",
  prs = 1,
  e = 0.1
)
}
\arguments{
\item{X}{Design matrix}

\item{Zcols}{Vector of random effect indices in design matrix}

\item{Y}{Outcome variable}

\item{ni}{Vector indicating the repeated observations of each subject}

\item{Th0}{Output from \code{lme_mass_fit_init}}

\item{Rgs}{Output from \code{lme_mass_RgGrow}}

\item{Surf}{Spherical surface}

\item{fname}{(currently ignored)}

\item{Dtype}{Distance (default: euclidean)}

\item{sptm}{Spatial measure (default: exp)}

\item{prs}{Number of cores for parallel computing (default: 1)}

\item{e}{Tolerance (default: 10^-1)}
}
\value{
This function returns a list of lists, with entries st (=status) and stats,
with the following entries: Bhat, CovBhat, phisqhat, Dhat, Zcols, invEI,
Pth, Qthth, lreml
}
\description{
Title
}
\examples{
\dontrun{fitRgw <- lme_mass_fit_Rgw(X, Zcols, Y, ni, Th0, Rgs, Surf)}
}
